+++ pakyow-presenter/spec/features/presenter_version_spec.rb
@@ -2,10 +2,10 @@ RSpec.describe "presenter version" do
   include_context "app"
 
   it "builds the presenter version" do
-    expect(Pakyow.apps.first.config.presenter.version).to eq("df628a7ca80870bb8c943cb895e6fe50ee83f3a5")
+    expect(Pakyow.apps.first.config.presenter.version).to eq("ed6c73c00503ae201c01e0ffc42b14a8df473054")
   end
 
   it "rebuilds the app version to include the presenter version" do
-    expect(Pakyow.apps.first.config.version).to eq("01cfb83ca46a1af1a4b8a73629762cf6b9c74964")
+    expect(Pakyow.apps.first.config.version).to eq("7f63c5274c46760e956d0e95f7aac3d0eef16456")
   end
 end