+++ pakyow-reflection/lib/pakyow/application/behavior/reflection/reflecting.rb
@@ -38,12 +38,48 @@ module Pakyow
               # Build the endpoints.
               #
               builders[:endpoints].build(@mirror.endpoints)
+            end
+
+            after "boot" do
+              @mirror.endpoints.each do |endpoint|
+                endpoint.exposures.each do |exposure|
+                  define_children_for_endpoint_context(exposure)
+                end
+              end
 
               # Cleanup.
               #
               unless Pakyow.env?(:test)
                 @mirror.scopes.each(&:cleanup)
                 @mirror.endpoints.each(&:cleanup)
+              end
+            end
+          end
+
+          private
+
+          def define_children_for_endpoint_context(exposure)
+            exposure.nodes.each do |exposure_node|
+              exposure_node.find_significant_nodes(:endpoint).each do |endpoint_node|
+                if app_endpoint = endpoints.find(name: endpoint_node.label(:endpoint))
+                  app_endpoint.params.each do |param|
+                    data.send(exposure.scope.plural_name).source.class.associations.values.flatten.each do |association|
+                      association_key_prefix = "#{association.name}_"
+                      if param.to_s.start_with?(association_key_prefix)
+                        scope = @mirror.scopes.find { |scope|
+                          scope.plural_name == association.associated_source_name
+                        } || Pakyow::Reflection::Scope.new(association.name)
+
+                        Pakyow::Reflection::Exposure.new(
+                          scope: scope,
+                          nodes: [endpoint_node],
+                          parent: exposure,
+                          binding: exposure.binding
+                        )
+                      end
+                    end
+                  end
+                end
               end
             end
           end