+++ pakyow-reflection/lib/pakyow/reflection/builders/helpers/controller.rb
@@ -224,9 +224,6 @@ module Pakyow
               )
             )
           end
-
-
-
         end
       end
     end