+++ pakyow-presenter/CHANGELOG.md
@@ -2,6 +2,7 @@
 
   * [changed] Prototype bar is now devtools, available in development and prototype modes
   * [fixed] Incorrect renders when a collapsed string doc is transformed
+  * [changed] Endpoints now pull param values from related objects
 
 # 1.0