+++ pakyow-reflection/CHANGELOG.md
@@ -1,5 +1,7 @@
 # UNRELEASED
 
+  * [added] Include associated data for endpoint building
+
 # 1.0
 
   * Hello, Web